+++ pcs/common/reports/messages.py
@@ -4534,6 +4534,60 @@ class SbdNotInstalled(ReportItemMessage):
 
 
 @dataclass(frozen=True)
+class SbdWithDeviceCannotSetWatchdogTimeout(ReportItemMessage):
+    """
+    Can't set stonith-watchdog-timeout When using SBD with device
+    """
+    _code = codes.SBD_WITH_DEVICE_CANNOT_SET_WATCHDOG_TIMEOUT
+
+    @property
+    def message(self) -> str:
+        return "stonith-watchdog-timeout can only be set to 0 while sbd is using devices"
+
+
+@dataclass(frozen=True)
+class SbdNotUsedCannotSetWatchdogTimeout(ReportItemMessage):
+    """
+    Can't set stonith-watchdog-timeout When using SBD with device
+    """
+    _code = codes.SBD_NOT_USED_CANNOT_SET_WATCHDOG_TIMEOUT
+
+    @property
+    def message(self) -> str:
+        return "Unable to set stonith-watchdog-timeout while sbd is disabled"
+
+
+@dataclass(frozen=True)
+class StonithWatchdogTimeoutInvalid(ReportItemMessage):
+    """
+    stonith-watchdog-timeout is not integer
+    """
+    _code = codes.STONITH_WATCHDOG_TIMEOUT_INVALID
+
+    @property
+    def message(self) -> str:
+        return "The value of stonith-watchdog-timeout must be integer"
+
+
+@dataclass(frozen=True)
+class StonithWatchdogTimeoutTooSmall(ReportItemMessage):
+    """
+    The value of stonith-watchdog-timeout is too small
+
+    local_sbd_watchdog_timeout -- local sbd watchdog timeout
+    """
+    local_sbd_watchdog_timeout: int
+    _code = codes.STONITH_WATCHDOG_TIMEOUT_TOO_SMALL
+
+    @property
+    def message(self) -> str:
+        return (
+            "The stonith-watchdog-timeout must be greater than "
+            f"{self.local_sbd_watchdog_timeout}"
+        )
+
+
+@dataclass(frozen=True)
 class WatchdogNotFound(ReportItemMessage):
     """
     Watchdog doesn't exist on specified node